refactor: domain directory structure
